requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
python-dotenv>=1.0.1
python-docx>=1.1.2
pymupdf>=1.24.0
pydantic>=2.7.0
pandas>=2.2.0
openpyxl>=3.1.2
rich>=13.7.1
py-xmind16>=0.1.0
streamlit>=1.33.0
httpx>=0.27.0
trafilatura>=1.8.0
beautifulsoup4>=4.12.0
